+++ lib/common/presentation/router/app_router.dart
@@ -61,6 +61,10 @@ class AppRouter extends RootStackRouter {
           guards: [AuthGuard()],
         ),
         MaterialRoute(
+          page: DeliveryDetailRoute.page,
+          guards: [AuthGuard()],
+        ),
+        MaterialRoute(
           page: ChangeActivePairRoute.page,
           guards: [AuthGuard()],
         ),