+++ lib/features/food/presentation/screens/history_screen.dart
@@ -1,9 +1,6 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:zachranobed/common/domain/utils/constants.dart';
 import 'package:zachranobed/common/domain/utils/date_time_utils.dart';
-import 'package:zachranobed/common/presentation/model/food_category.dart';
 import 'package:zachranobed/common/presentation/utils/build_context_extensions.dart';
 import 'package:zachranobed/common/presentation/utils/helper_service.dart';
 import 'package:zachranobed/common/presentation/utils/image_assets.dart';
@@ -14,14 +11,11 @@ import 'package:zachranobed/common/presentation/widget/page/loading_page.dart';
 import 'package:zachranobed/common/presentation/widget/screen_scaffold.dart';
 import 'package:zachranobed/common/presentation/widget/sectioned_list_view.dart';
 import 'package:zachranobed/common/presentation/widget/ui_app_bar.dart';
-import 'package:zachranobed/common/presentation/widget/ui_icon.dart';
-import 'package:zachranobed/common/presentation/widget/ui_meal_badge.dart';
-import 'package:zachranobed/common/presentation/widget/ui_meal_tile.dart';
-import 'package:zachranobed/features/food/domain/model/food_date_time.dart';
 import 'package:zachranobed/features/food/domain/model/offered_food.dart';
 import 'package:zachranobed/features/food/domain/usecase/get_history_paginated_use_case.dart';
 import 'package:zachranobed/features/food/presentation/model/food_allergen.dart';
 import 'package:zachranobed/features/food/presentation/widget/food_allergens_bottom_sheet.dart';
+import 'package:zachranobed/features/food/presentation/widget/meal_tile_factory.dart';
 
 /// A screen that displays the paginated history of offered food donations.
 ///
@@ -235,80 +229,8 @@ class _HistoryList extends StatelessWidget {
         controller: controller,
         physics: const AlwaysScrollableScrollPhysics(),
         entries: entries,
-        itemBuilder: (context, item) {
-          return _buildItem(context, item);
-        },
+        itemBuilder: (context, item) => MealTileFactory.buildMealTile(context, item),
       ),
-    );
-  }
-
-  Widget _buildItem(BuildContext context, OfferedFood item) {
-    return UiMealTile(
-      title: item.dishName,
-      quantityLabel: _formatQuantity(context, item),
-      badges: [
-        _buildCategoryBadge(context, item),
-        _buildAllergensBadge(context, item),
-        _buildDateBadge(context, item),
-      ],
-    );
-  }
-
-  String _formatQuantity(BuildContext context, OfferedFood item) {
-    if (item.numberOfServings != null) {
-      return context.l10n.commonServingsCount(item.numberOfServings!);
-    }
-    if (item.numberOfPackages != null) {
-      return context.l10n.foodInfoCountTemplate(item.numberOfPackages!);
-    }
-    return '';
-  }
-
-  UiMealBadge _buildCategoryBadge(BuildContext context, OfferedFood item) {
-    return switch (item.foodCategoryType) {
-      FoodCategoryType.warm => UiMealBadge(
-          icon: UiIconSpec.svg(ImageAssets.iconHot),
-          label: context.l10n.foodCategoryBadgeWarmTemplate(item.foodTemperature ?? Constants.foodTemperatureInitial),
-        ),
-      FoodCategoryType.cooled => UiMealBadge(
-          icon: UiIconSpec.svg(ImageAssets.iconCold),
-          label: context.l10n.foodCategoryBadgeCooled,
-        ),
-      FoodCategoryType.packaged => UiMealBadge(
-          icon: UiIconSpec.svg(ImageAssets.iconPack),
-          label: context.l10n.foodCategoryBadgePackaged,
-        ),
-      null => UiMealBadge(
-          icon: UiIconSpec.svg(ImageAssets.iconMeal),
-          label: item.foodCategory,
-        ),
-    };
-  }
-
-  UiMealBadge _buildAllergensBadge(BuildContext context, OfferedFood item) {
-    String allergensLabel;
-    if (listEquals(item.allergens, [FoodAllergen.noAllergensNumber])) {
-      allergensLabel = context.l10n.allergensNotPresent;
-    } else if (listEquals(item.allergens, [FoodAllergen.onPackageNumber])) {
-      allergensLabel = context.l10n.allergensOnPackage;
-    } else {
-      allergensLabel = item.allergens.join(', ');
-    }
-    return UiMealBadge(
-      icon: UiIconSpec.svg(ImageAssets.iconAllergens),
-      label: allergensLabel,
-    );
-  }
-
-  UiMealBadge _buildDateBadge(BuildContext context, OfferedFood item) {
-    final dateLabel = switch (item.consumeBy) {
-      FoodDateTimeSpecified(date: final date) => DateTimeUtils.formatDateTime(date, "d.M.y HH:mm"),
-      FoodDateTimeOnPackaging() => context.l10n.foodDateTimeLabelOnPackaging,
-    };
-
-    return UiMealBadge(
-      icon: UiIconSpec.svg(ImageAssets.iconCalendar),
-      label: dateLabel,
     );
   }
 }